refactor: atomic design에 맞게 위치 수정
